// frontend/src/api-client/types.gen.ts
export type ErrorResponse = {
  error: string;
  message: string;
};

export type Pagination = {
  count: number;
  currentPage: number;
  nextPage: number;
  pageSize: number;
  pages: number;
  previousPage: number;
};

export type Role = {
  createdAt: string;
  description: string;
  id: string;
  name: string;
  permissions: Array<string>;
  updatedAt: string;
};

export type Roles = Array<Role>;